# app/api/v1/endpoints/cierres.py
from fastapi import APIRouter, Depends
from sqlalchemy.ext.asyncio import AsyncSession
from app.core.dependencies import get_db, get_current_user
from app.schemas.cierre import CierreDiarioCreate, CierreDiarioRead

router = APIRouter()


@router.get("/", response_model=list[CierreDiarioRead])
async def list_cierres(
    db: AsyncSession = Depends(get_db),
    current_user: Usuario = Depends(get_current_user),
):
    raise NotImplementedError


@router.post("/", response_model=CierreDiarioRead, status_code=201)
async def create_cierre(
    cierre: CierreDiarioCreate,
    db: AsyncSession = Depends(get_db),
    current_user: Usuario = Depends(get_current_user),
):
    raise NotImplementedError